+++ video/out/cocoa_cb_common.swift
@@ -133,7 +133,6 @@ class CocoaCB: Common, EventSubscriber {
         switch outputCsp {
         case MAC_CSP_AUTO: return (false, colorSpace)
         case MAC_CSP_DISPLAY_P3: return (true, CGColorSpace(name: CGColorSpace.displayP3))
-        case MAC_CSP_DISPLAY_P3_HLG: return (true, CGColorSpace(name: CGColorSpace.displayP3_HLG))
         case MAC_CSP_DCI_P3: return (true, CGColorSpace(name: CGColorSpace.dcip3))
         case MAC_CSP_BT_2020: return (true, CGColorSpace(name: CGColorSpace.itur_2020))
         case MAC_CSP_BT_709: return (false, CGColorSpace(name: CGColorSpace.itur_709))
@@ -142,6 +141,13 @@ class CocoaCB: Common, EventSubscriber {
         case MAC_CSP_RGB_LINEAR: return (false, CGColorSpace(name: CGColorSpace.genericRGBLinear))
         case MAC_CSP_ADOBE: return (false, CGColorSpace(name: CGColorSpace.adobeRGB1998))
         default: break
+        }
+
+        if #available(macOS 10.14.6, *) {
+            switch outputCsp {
+            case MAC_CSP_DISPLAY_P3_HLG: return (true, CGColorSpace(name: CGColorSpace.displayP3_HLG))
+            default: break
+            }
         }
 
 #if HAVE_MACOS_10_15_4_FEATURES